CVS_SILENT: dont need uint, only int
svn path=/trunk/kdeextragear-1/amarok/; revision=405337

+++ src/playlist.cpp
@@ -1818,12 +1818,11 @@ Playlist::removeSelectedItems() //SLOT
 
     if( AmarokConfig::partyMode() )
     {
-        uint tracks = AmarokConfig::partyUpcomingCount() + AmarokConfig::partyPreviousCount() + 1; //+1 for current track
-
-        if( (int)tracks > childCount() )
+        int tracks = AmarokConfig::partyUpcomingCount() + AmarokConfig::partyPreviousCount() + 1; //+1 for current track
+        if( tracks > childCount() )
             addSpecialTracks( tracks - childCount(), AmarokConfig::partyType() );
-
     }
+
     updateNextPrev();
     //NOTE no need to emit childCountChanged(), removeItem() does that for us
 }
